requirements.txt
attrs==22.2.0
cstruct==5.2
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
posix-ipc==1.1.1
pytest==7.2.1
PyYAML==6.0
tomli==2.0.1
